requirements.txt
# AI_GF Project Requirements

# Aiogram - фреймворк для создания Telegram ботов
aiogram==3.13.1

# SQLAlchemy - ORM для работы с базами данных
sqlalchemy==2.0.36

# Дополнительные библиотеки
aiosqlite==0.20.0  # Асинхронный драйвер для SQLite
asyncpg==0.30.0  # Асинхронный драйвер для PostgreSQL
psycopg2-binary==2.9.10  # PostgreSQL адаптер (бинарная версия)
greenlet==3.1.1  # Требуется для асинхронного SQLAlchemy
python-dotenv==1.0.1  # Для работы с переменными окружения (.env файлы)
alembic==1.14.0  # Миграции базы данных

# Redis для хранения истории чатов
redis==5.0.1

# RabbitMQ для очередей сообщений
pika==1.3.2

# HTTP клиент для запросов к LLM API
aiohttp==3.10.11

# Векторный поиск и эмбеддинги
sentence-transformers==5.1.1
chromadb==1.1.1
numpy==2.3.3

# Дополнительные зависимости для векторного поиска
torch==2.8.0
transformers==4.57.1
scikit-learn==1.7.2
scipy==1.16.2

